# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация API к приложению достопримечательностей",
        "description": "Авторизация происходит через OAuth2, описание и ссылки на методы надоятся в Authorize",
        "contact": {
            "email": "sergei.klimovich.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://travel.madskill.ru",
            "description": "API сервер"
        }
    ],
    "paths": {
        "/api/populate": {
            "get": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Популярные Достопримечательности",
                "description": "Популярные Достопримечательности",
                "operationId": "populate",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Populate"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recommendations": {
            "get": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Рекомендуемые Достопримечательности",
                "description": "Рекомендуемые Достопримечательности",
                "operationId": "recommendations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Populate"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/search": {
            "get": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Поиск Достопримечательностей",
                "description": "Поиск Достопримечательностей",
                "operationId": "search",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Город",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Populate"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/attraction": {
            "get": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Одна Достопримечательность",
                "description": "Одна Достопримечательность",
                "operationId": "attraction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SingleAttraction"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rating": {
            "get": {
                "tags": [
                    "Рейтинг"
                ],
                "summary": "Получить Рейтинг",
                "description": "Получить Рейтинг",
                "operationId": "rating",
                "parameters": [
                    {
                        "name": "attraction_id",
                        "in": "query",
                        "description": "id достопримечательности",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Raiting"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Рейтинг"
                ],
                "summary": "Изменить Рейтинг",
                "description": "Изменить Рейтинг",
                "operationId": "postrating",
                "parameters": [
                    {
                        "name": "attraction_id",
                        "in": "query",
                        "description": "id достопримечательности",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "like",
                        "in": "query",
                        "description": "Положительная эмоция",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dislike",
                        "in": "query",
                        "description": "Отрицательная эмоция",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Raiting"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Photo": {
                "title": "Photo",
                "description": "Модель фотографий",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "attraction_id": {
                        "title": "attraction_id",
                        "description": "attraction_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "image": {
                        "title": "image",
                        "description": "Название",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    }
                },
                "type": "object"
            },
            "Populate": {
                "title": "Populate",
                "description": "Модель достопримечательности",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Название",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "description": {
                        "title": "description",
                        "description": "Описание",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "image": {
                        "title": "image",
                        "description": "Ссылка на фото",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "city": {
                        "title": "city",
                        "description": "Город",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    }
                },
                "type": "object"
            },
...
            "Raiting": {
                "title": "Raiting",
                "description": "Модель рейтингаи",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "attraction_id": {
                        "title": "attraction_id",
                        "description": "attraction_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "dislike": {
                        "title": "dislike",
                        "description": "dislike",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "like": {
                        "title": "like",
                        "description": "like",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "SingleAttraction": {
                "title": "SingleAttraction",
                "description": "Модель одной достопримечательности",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Название",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "description": {
                        "title": "description",
                        "description": "Описание",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "city": {
                        "title": "city",
                        "description": "Город",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "photo": {
                        "title": "photo",
                        "description": "photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Внимание OAuth2 scopes none, grant type code",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://travel.madskill.ru/oauth/authorize",
                        "tokenUrl": "https://travel.madskill.ru/oauth/token",
                        "refreshUrl": "https://travel.madskill.ru/token/refresh",
                        "redirectUrl": "https://travel.madskill.ru/auth/callback",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}